requirements.txt
# PDF and Image Processing
docling>=2.0.0
docling-core>=2.0.0
PyMuPDF>=1.23.0
pdf2image>=1.16.3
Pillow>=10.0.0

# OpenAI API
openai>=1.0.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
pydantic>=2.0.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.65.0
requests>=2.31.0

# JSON Schema Validation
jsonschema>=4.0.0

# PowerPoint Generation
python-pptx>=0.6.21 